Evaluate binary ops first
This is blatantly premature optimization.  Absolutely no need for it other
than the ridiculous thought that binary ops are more common and we save
some processing time by evaluating the if earlier.  The amount of time
conceivably saved by this is far less significant than the amount of time
I have already wasted typing this.  It's funny the things we do....