Terminate a numeric imput on [
This whole termination of input buffer needs to be cleaned up.  We
should probably just pop a value of the stack to handle 'x', but
I'm not sure how we should handle input like '4[1 1+]x'.  Intuitively,
the "1 1+" should be executed leaving 4 and 2 on the stack, which is
what we do now, and '4[1 1+x][2 2+x]0x' will execute the bottom
string in the stack, and not leave 0 on the stack.