Pass size as distinct argument to xrealloc, and reassign an iterator
This will catch a large class of bugs.  I have never before been a fan
of this style, reminescent of calloc, but I think this is a good thing.
I did lose a good 20 minutes yesterday debugging a failure to include
the size in tac.c, and while implementing this caught the previously
unnoticed error at line 105 of read-indexed.c.