Introduce memory issue
This was the first attempt to handle the 2nd half of the problem, and introduces
a memory bug I have not yet (and probably never will) track down.  In particular,
the call to find_type_in_dir at line 57 is getting messed up.  The value of i
increments, but the value of dir in the function is always 1.  It would seem
the stack is getting crushed, but I do not see the error.  I am going to reimplement
using a single dimensional buffer to hold the data and use fgetc for input
and make the problem go away.  Just wanted to record this version for posterity
in case I ever want to debug it.