Update dynamic-array.c
Change the name from "pop" to "next" so it does not feel like a stack.
Use a data pointer so calls to push and next can be interleaved.
Use unsigned char so this will work with non-text files. (eg, -1
in the data will not be converted to (int)-1 and treated like EOF.