Import interface.Example
It is not clear to me what the leading capitalization does.  Using
the keyword "interface" as the module name leads to a syntax error,
and the leading capialization skirts the error.  However, since go
has made the phenomenally stupid design decision to make visibility
of symbols depend on the naming scheme, there may be implications
of this capital letter that are not yet understood by me.