Remove unnecessary initialization
The initializations are only useful if we intend
to allow random access into the array.  Since we
are monotonically incrementing the index, it is not
necessary.  Since the purpose of this code is strictly
educational, it seems that clarity is better than
flexibility.  This may be a terrible decision.  It would
be better to move the line_count increment into this
function to encapsulate that....but this is just a toy
project.