Do not bother with NDEBUG and O2 in bench
See comment in most recent commit.  We need these applied when the library is
built, not when the wrapper is built.  But this requires flags to be set too early
unless we kludge the benchmark in the source for the libary, which is unappealing
and is a lie.  The only way benchmarks can have any meaning is if they are run
against an installed library, so they really need to be in a separate project.