Tweak hashmap_clear to resize
This needs some thought.  git bisect shows the recent errors
coming from 140c31b3043a742417e65963907, but this is not
caused by a coding error.  Rather, by resizing the hashmap
to be small, we are now seeing collisions in the test that
were not there before.  Need to fix the test