Take seeds out of hash method
This hash_method structure was created to help clarify the interface, and
is overly complicated.  Most hash functions will not need or want two
seeds (eg the provided murmur implementation ignores the second seed),
and even putting a void * in the hashmap as we do here seems overkill, but
we do it this way to minimize churn.