[*BROKEN*] Attempt to intercept malloc
Using custom allocators is a bad idea for several reasons:

this is a toy, and should be simple
we can intercept malloc

Sadly, the implementation here is not working, and I suspect it is
a MacOS security feature that is preventing us from actually
intercepting.  Just want to get this commit somewhere other
than in a stash on my local machine.