Do not use '%d' as a format specifier
If the string representation of a value cannot fit in an int (eg,
if the value is over INT_MAX), the behavior is undefined.  This issue
can be mitigated with '%4d' or similar, but this greatly restricts the
actual range of usable values.  Doing something this simple with scanf
is surprisingly (absurdly) complicated.  sigh.  It is good to have this
example somewhere, I suppose.