Remove explicit references to MAKEFLAGS
This is broken.  MAKEFLAGS get set to values without preceding hyphen,
so passing it explicitly causes a failure.  make ought to be taking
options directly from MAKEFLAGS in its environment, so there does not
seem to be any need to pass it explicitly.  I do not know that i was
thinking when I added that in the first place.  It may be worth
tracking it down to see if I made a comment, but most likely
it was just an error.