Move dotfiles to docker/linux
This is not clean.  I should be pulling the dotfiles out of
the dotfiles git repo instead of duplicating them here, but
that repo does not really contain minimal dotfiles and there
just does not seem to be a clean way to import without
referencing the url on a public repo, and that seems kludgy.
Perhaps the default workflow should be to build the image
with minimal dotfiles but expect the user to mount their
local home directory in the container and not use them at
all.