Rename calc-test.in calc-test
I am preparing to setup a bazel build, and am not
yet sure how to do feature testing.  The calc-test
was using autoconf to determine if the c library's
printf supported ' as a digit separator in format
strings (search for HAVE_COMMA_FORMAT in the recently
deleted configure.ac), and the substitution in the
newly named calc-test will suppress that test on all
platforms.

TODO: figure out feature testing in bazel.

Signed-off-by: William Pursell <william.r.pursell@gmail.com>