Move the libraries into the root module
The attempt to include headers from a sibling package worked,
but the build was generating the warning:

in includes attribute of cc_library rule //lib:ringbuf: ../include resolves to include not below the relative path of its package lib. This will be an error in the future

Putting the libraries in the root module is absolutely the wrong
way to resolve this, but I am just playing around with bazel
and this is educational.

Signed-off-by: William Pursell <william.r.pursell@gmail.com>