Discard error via "catch unreachable"
This is a code smell.  OTOH, ignoring the error is just as bad, so maybe
this is a case of the zig philosophy doing exactly the right thing.  The
issue here is that we cannot "defer try", and the language is forcing
us to explicitly deal with the error, which is absolutely correct behavior.
And we are skirting the issue entirely, and the language is letting us
do so without too much friction, and it is clear that there is a potential
footgun here.