Honor OOOOK
I am noticing that ~/.bash-history is horribly out of order, and
this script seems like it could be useful for filtering.  If we
build another script to sort the files, we could actually have
rather robust merging ability.  For now, though, let's just
allow files to be mal-ordered, since this lets me filter commands.