Append filename to non-annotated timestamps
The idea is that we want to have the hostname, but many bash history
files are totally uncommented.  When we gather the history from a remote
host, we expect to put it in a file that contains the hostname, and
this will add that data to the timestamps of the merged file.