Refactor
The while loop in init is redundant, since
we have that loop in next().  I am not really
sure how to behave if the input file does not
begin with a timestamp.  It seems like that
ought to be treated as an error, but I do not
anticipate what the use cases are.  Seems like
it might be useful to allow partial files that
do not begin as expected.  Really need to implement
a test suite, though.