Break out class method definitions
Having all function definitions in the top level class definition
leads to unmaintainable garbage, with class definitions that span
hundreds of lines.  It would be much better if we could create
multi-line lambdas or do something like "def foo::__init__(...",
but this is the closest I can currently think of for a reasonable
method.