Add some test cases
I am fairly sure I see the approach to take.  Sort
the heights.  The tallest must either be one of the
edges or it will be in the middle.  If an edge
is the higher (or equal) boundary, there cannot be
a higher piece on the other side.  (eg, if point
(n, m) is the right boundary and m > h (where h
is the height of the left boundary), then no value
to the right of n can have height greater than m.

Basically, we will find the 2 highest points and call
that the putative solution, then check the coordinates
in order of descending height; at each step we just need
to compare the area of the box with the previous and
this becomes a linear search moving outward from the
highest point.