Implement linear search
Still crazy slow; could remove the reallocs and think a bit
more about the algorithm, but this passes the test suite.

low hanging fruit: filter data points when building the
queue.  Undoubtedly there are easy filters to throw out
points that are not going to be relevant.  (eg, if
a short value lies between to larger values, it can probably
be ignored.)