Exit with 1, not 5
I just read that the default return status is 5.  That is
incredibly stupid, so we explicitly return 1.  Note
that you can use "halt" to exit 0, but you cannot
pass a value to halt so we need to discard data
as in the example to avoid spewing the input onto
the error stream.