Attempt to fix display with setlinebuf
Hmm.  This is working well, but if the user hits "return" and
injects a newline into the terminal, the cursor will jump
all over until the next SIGINT, at which point things start
looking good again.  (On a terminal on my mac laptop).  This
is a (failed) attempt to fix that issue.  Perhaps we should
just bring back the kludge and make the cursor invisible, but
this is curious behavior.