Remove obsolete comment
We are now using tput via system to hide the cursor,
and this is a great example of how easily comments
can go stale.  This is a single developer over the course
of a few days (18 hours!?) forgetting to update a comment.

Comments should *never* describe what code is doing.
There should be one source of truth, and comments are
not it.  Comments are for explaining why a choice
was made (and illuminating gotchas) and must remain
minimal!