Use "stat -f" if -c is not available
Sigh.  Standardization for tools is absurd.  My guess is that I was
using stat -c from gnu coreutils and now I have a version that does
not recognize it.  This is a bit kludgy but for now we seem to
get the right behavior from stat -f.  TODO: see if stat -f works
on Linux with coreutils.

This check is necessary, since the .gdca file gets modified during
processing, so the test will fail if we try to compare it to
its original form.