Suppress output
If the cmp fails, it is probably because the file
was recently modified and we are about to ignore
the test, so we do not care about the output.

Even if the test correctly indicates a problem,
the output of cmp is not helpful, so it is fine
to suppress it.

The actual case that is happening is the .gcda
file, which gets modified during the execution
of the test and hence the output of the test
does not match the original file.