Read 128 char at a time
This is pretty pointless, but feels aesthetically better.
There is no need to check for newline, as we are already
assuming data is coming from a terminal that is buffering
until user hits <enter>, and if the user does hit <enter>
twice quickly enough that we are still in this function
we probably only want to print one delta anyway.