Clean up read()
The &b in the read is odd, and I believe we get
the desired behavior with this new flow.  The
idea here is just that we want to make sure that
the user interaction only causes a single
line to print and are trying to guard against
double taps.

lang: c