Use hash table to lookup functions
This is a complete hack.  At the moment, we rely
on the the lack of collisions in the hash table
which we can do since the number of functions is
small and known at compile time.  This will not
scale, but the whole point of this code is to be
a silly little toy to enable procrastination, so
this is an acceptable trade-off.

lang: c