Revert previous 2 commits (storing ull and ld)
Eventually, it probably makes sense to do something like
this, since we should allow bit-operators like & and |, and
probably want to consider making ^ a bit operator instead
of exponentiation (we can use pow), but doing so
requires a bit more thought and architectural changes
than I am willing to do right now.  Remember, this
is just a toy and is not worth spending any time on.

lang: c