If top of stack is not an integer, use most recent register
This seems like a more natural interface, but
it is not clear if a warning should be emitted.
When a user does something like: "2.3[%Lf]Fp", it
seems clear that they want to use the most recent
register as the format string, and attempting to use
2.3 as an index into the register stack is silly.
On the other hand, "3[%Lf]Fp" will still attempt
to use 3 as an index into the register stack, which
seems inconsistent.  Perhaps we should use a state
flag of some sort and only pop an index from the
stack if the register was entered after the last
value on the stack.

lang: c