Replace "Rigidly" with "Vigorously"
This is important.  Formatting guidelines and linters should never
be "rigid".  One thing we really need is a better flexible linter,
and I regret ever using the word "Rigidly".  We should never
impose a linter on a code base that already exists, since that
invariably leads to many churn commits that do nothing but update
the style, and that would be greatly reduced with better linters.

Not sure if I should add a bullet point about not imposing
a linter after the code base is established, since what we
really need is better linters and more discipline around avoiding
churn commits.

lang: doc