Rename function
Hmmm.  This is a bit of a mess.  Currently, strings like
+56+34+345 ... will need to be completely accumulated
in the fixed sized buffer, and the processing of those +
signs to decide that they are part of a number and not
a binary operator is being deferred for too long.  We
need to push values onto the stack sooner.  Probably should
handle +- specially in the main loop.

lang: c