Un-factor push_value()
This is a mess.  We don't want to include parsing logic here,
but since + and - are each binary operators and numeric_tokens,
we sort of have to.  Currently, we keep all of the logic in
strtold, but this means we are potentially popping more off
the stack than we need to.  It seems that unless the previous
character was 'e' or 'E', any instance of '+' or '-' should
be treated as a binary operator, so we can return early.  But just
checking that and breaking out of the loop is not enough,
and we still need to keep things as is.  In particular, it
is stupid that on input like +8+7+6+5, we put everything
into a buffer which can easily overflow. And on input like
"8e+++++---++8375++---+" we are popping everything off the
stack, then deteting the syntax error, and then pushing
everything back on the stack.  We should pop as little
as possible.

lang: c