Rename pop_value -> wrap_pop
This function is merely a wrapper around stack_pop that emits
an error message.  I want to implement a pop_value function
that takes the state struct as an argument and stores the
value popped in an alternate stack, allowing the user full
access to the memory stack to recall previous values.  I think
that would be useful functionality.

lang: c