Add test which duplicates an item in memory
We are about to introduce a code change which attempts to ensure
that stack entries are not duplicated in memory.  This does
not prevent values from being the same in multiple entries
in the memory stack, but it will ensure that a single entry
(ie, one which has been moved to the stack with the m command)
will not be on the memory stack more than once.

lang: c