Print values based on their type
This is incomplete, and changes the semantics of the k command.
0k would set the state type to integer, but we want instead
to use i to set an input type and use strtoll for input, and assign
a type at that point.  Also, we need to check the input types
when calling any functions and potentially covert to double.

Still a work in progress.

lang: c