Start requiring a space before operators
I think this can work.  Need to think about it a bit, but I believe
the implementation will be trivial if we simply require a space
before an operator.  Then "3 -5 +" and "3 -5e+2 +" will work.
Basically, we collect symbols until whitespace and then try
to parse a number.  If strtold shows the token is not a number,
then we treat each symbol as an operator.  And we treat "3 -5+"
as a syntax error.