Push value when +/- seen
Currently, we are accumlating all of the input into a small
fixed sized buffer, so strings like "+2+2+2...." are not parsed
until whitespace or similar is seen.  With this change, we
start parsing numbers before we see any whitesapce, so we are
less likely to overflow the buffer.  This not an important change,
but feels aesthetically appealing.  Mostly I suppose I'm still
procrastinating on cleaning up the handling to push_value,
which is quite ugly.