Deprecate F
This change adds some code to parse the user specified format
string purely so that we can minimize the changes in the test
suite.  I am unsure if we want to give the option to add +
or leading zeros to the format string.  Need to focus on
changing the test suite to match desired behavior to make
that decision.

lang: c