Always use lld for integer types
We should probably add a separate format string to allow a specification
for integer types, but applying the width modifier by default
is annoying, as integers like "5" are printed as "005" using the
default format.  But I like the default format for long double.

We could play around with different options, but this seems
reasonable for the moment.

lang: c