Use frexp to print human readable values
This is a bit confusing.  frexp returns a value
in [.5, 1), and the exponent is one greater than
the actual stored exponent.  eg, for 1.0, the
(stored) mantissa is zero, and the exponent is
1027.  With bias, that exponent becomes zero.
The stored value is 1 + mantissa (hidden bit),
so the double representing 1.0 is 1 * 2 ^ (1027 - 1027),
but frexp breaks it into (.5) * 2 ^ 1.  It would
be more natural for frexp to return a value in
the interval [1,2).